% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.clean_club_name}
\alias{.clean_club_name}
\title{Internal function to clean club name by removing abbreviation}
\usage{
.clean_club_name(club_names)
}
\arguments{
\item{club_names}{Vector of club names}
}
\value{
Vector of cleaned club names
}
\description{
Internal function to clean club name by removing abbreviation
}
\keyword{internal}
