% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/races.R
\name{.extract_all_races}
\alias{.extract_all_races}
\title{Internal function to extract all race tables}
\usage{
.extract_all_races(html_content)
}
\arguments{
\item{html_content}{Parsed HTML content}
}
\value{
List of race data frames
}
\description{
Internal function to extract all race tables
}
\keyword{internal}
