% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.extract_club_abbreviation}
\alias{.extract_club_abbreviation}
\title{Internal function to extract club abbreviation from name}
\usage{
.extract_club_abbreviation(club_names)
}
\arguments{
\item{club_names}{Vector of club names}
}
\value{
Vector of abbreviations
}
\description{
Internal function to extract club abbreviation from name
}
\keyword{internal}
