% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/races.R
\name{.extract_club_info_from_table}
\alias{.extract_club_info_from_table}
\title{Internal function to extract club information from a race table}
\usage{
.extract_club_info_from_table(table, racers_df)
}
\arguments{
\item{table}{HTML table element}

\item{racers_df}{Racers data frame for reference}
}
\value{
List of club information by racer name
}
\description{
Internal function to extract club information from a race table
}
\keyword{internal}
