% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/races.R
\name{.extract_racers}
\alias{.extract_racers}
\title{Internal function to extract unique racers}
\usage{
.extract_racers(html_content)
}
\arguments{
\item{html_content}{Parsed HTML content}
}
\value{
Data frame with racer information
}
\description{
Internal function to extract unique racers
}
\keyword{internal}
