% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/races.R
\name{.parse_race_table}
\alias{.parse_race_table}
\title{Internal function to parse a single race table}
\usage{
.parse_race_table(table)
}
\arguments{
\item{table}{HTML table element}
}
\value{
Data frame with race results
}
\description{
Internal function to parse a single race table
}
\keyword{internal}
