% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/races.R
\name{.preprocess_race_table}
\alias{.preprocess_race_table}
\title{Internal function to pre-process race table HTML}
\usage{
.preprocess_race_table(table)
}
\arguments{
\item{table}{HTML table element}
}
\value{
Pre-processed HTML table element
}
\description{
Applies the pre-processing rules for headers and body rows before
using rvest::html_table()
}
\keyword{internal}
