% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/races.R
\name{.process_table_body}
\alias{.process_table_body}
\title{Internal function to process table body}
\usage{
.process_table_body(table)
}
\arguments{
\item{table}{HTML table element}
}
\value{
Table with processed body
}
\description{
Rule 6.1: If body rows contain nesting, choose the element's inner text
Rule 6.2: If body row contains nesting and element's inner text is blank
and children have the 'points-' class, ignore this element and all child elements
Rule 6.3: If body rows contain attribute 'class=win_for_{n}', capture the element attribute text
Rule 6.4: Insert captured element attribute text into the empty element
}
\keyword{internal}
