% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{extract_tables}
\alias{extract_tables}
\title{Extract tables from ski results HTML files}
\usage{
extract_tables(file_path)
}
\arguments{
\item{file_path}{Path to the HTML file}
}
\value{
List of dataframes containing the extracted tables
}
\description{
Extract tables from ski results HTML files
}
