% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{process_raceEvent_files}
\alias{process_raceEvent_files}
\title{Process all HTML files in the package's extdata directory}
\usage{
process_raceEvent_files()
}
\value{
List of lists containing tables from each HTML file
}
\description{
Process all HTML files in the package's extdata directory
}
